test(doctor): consolidate startup host-link regression
